// Project2/Main/src/webextension/src/interfaces/ApiResponseModel/IReceiveConversation.ts
export interface IReceiveConversation {
    fullName: string;
    alias: string;
    roleChange: string;
    fyManagerChange: string;
    scriptLink: string;
    sendingConversationStatus: string;
    receiveConversationStatus: string;
    fyManagerAlias: string;
    cyManagerAlias: string;
    edmValidation: string;
    fyOrg: string;
    fyRoleSummary: string;
    fyQ1: string;
    fyQ2: string;
    fyIncentivePlan: string;
    fyCostCenter: string;
    cyOrg: string;
    cyRoleSummary: string;
    cyQ1: string;
    cyQ2: string;
    cyCareerStage: string;
    cyIncentivePlan: string;
    cyCostCenter: string;
    cyManagerFullName:string;
    hasActiveFutureManagerRequest:boolean;
    futureManagerRequestSubmittedBy:string;
    taxonomyCorrectionRequestSubmittedBy:string;
    hasActiveTaxonomyCorrectionRequest:boolean;
    fyManagerFullName:string;
    fyCareerStage:string;
    isEmployeeRecordApproved: boolean;
  }